auth-service/requirements.txt
flask==2.0.1
pyjwt==2.1.0
bcrypt==3.2.0
